chore(core): move backend protocols + connector key into core (#1922)
Move the backend Protocol surface (CodeHostBackend / MessagingBackend /
CIService / PullRequestSpec / ReviewState / ...) from teatree.backends.protocols
to teatree.core.backend_protocols, and the GRANTED_SCOPES_KEY connector key to
teatree.core.connector_keys, so the domain layer owns the abstractions it
consumes. teatree.backends.protocols stays a re-export shim and slack_scopes
re-imports the key, keeping every existing consumer working over the allowed
backends -> core direction.

Re-point all teatree.core (non-management) consumers to the core path and swap
reply_transport TYPE_CHECKING concrete-class hints to the core Protocols
(GitLabAPI -> a local structural GitLabNoteClient Protocol). No runtime
behaviour change; the Protocol surface is identical.

Part of breaking the core <-> backends import cycle.
