+++ src/teatree/core/backend_factory.py
@@ -13,7 +13,7 @@ from django.core.exceptions import ImproperlyConfigured
 
 from teatree.backends.loader import get_ci_service, get_code_host, get_code_hosts, get_messaging
 from teatree.backends.loader import reset_backend_caches as _reset_loader_caches
-from teatree.backends.protocols import CIService, CodeHostBackend, MessagingBackend
+from teatree.core.backend_protocols import CIService, CodeHostBackend, MessagingBackend
 from teatree.core.overlay import OverlayBase
 from teatree.core.overlay_loader import get_all_overlays, get_overlay
 from teatree.paths import find_overlay_db
